+++ src/System.Collections.Concurrent/tests/ConcurrentBagTests.cs
@@ -13,7 +13,7 @@ using Xunit;
 namespace System.Collections.Concurrent.Tests
 {
     /// <summary>The class that contains the unit tests of the LazyInit.</summary>
-    public class ConcurrentBagTests : IEnumerable_Generic_Tests<int>
+    public partial class ConcurrentBagTests : IEnumerable_Generic_Tests<int>
     {
         protected override IEnumerable<ModifyEnumerable> ModifyEnumerables => new List<ModifyEnumerable>();
         protected override IEnumerable<int> GenericIEnumerableFactory(int count) => new ConcurrentBag<int>(Enumerable.Range(0, count));
@@ -412,118 +412,6 @@ namespace System.Collections.Concurrent.Tests
         {
             DebuggerAttributes.ValidateDebuggerDisplayReferences(new ConcurrentBag<int>());
             DebuggerAttributes.ValidateDebuggerTypeProxyProperties(new ConcurrentBag<int>());
-        }
-
-        [Fact]
-        public static void RTest11_Clear()
-        {
-            RTest11_Clear(1, 10);
-            RTest11_Clear(3, 100);
-            RTest11_Clear(8, 1000);
-        }
-
-        public static void RTest11_Clear(int threadsCount, int itemsPerThread)
-        {
-            int addFailures = 0;
-            int peekFailures = 0;
-            int takeFailures = 0;
-            int clearFailures = 0;
-            int arrayFailures = 0;
-            ConcurrentBag<int> bag = new ConcurrentBag<int>();
-
-            Task[] addThreads = new Task[threadsCount];
-            Task[] peekThreads = new Task[threadsCount];
-            Task[] takeThreads = new Task[threadsCount];
-            Task[] clearThreads = new Task[threadsCount];
-            Task[] arrayThreads = new Task[threadsCount];
-
-            for (int i = 0; i < threadsCount; i++)
-            {
-                addThreads[i] = Task.Run(() =>
-                {
-                    for (int j = 0; j < itemsPerThread; j++)
-                    {
-                        try
-                        {
-                            bag.Add(j);
-                        }
-                        catch
-                        {
-                            Interlocked.Increment(ref addFailures);
-                        }
-                    }
-                });
-                peekThreads[i] = Task.Run(() =>
-                {
-                    for (int j = 0; j < itemsPerThread; j++)
-                    {
-                        try
-                        {
-                            int item;
-                            bag.TryPeek(out item);
-                        }
-                        catch
-                        {
-                            Interlocked.Increment(ref peekFailures);
-                        }
-                    }
-                });
-                takeThreads[i] = Task.Run(() =>
-                {
-                    for (int j = 0; j < itemsPerThread; j++)
-                    {
-                        try
-                        {
-                            int item;
-                            bag.TryTake(out item);
-                        }
-                        catch
-                        {
-                            Interlocked.Increment(ref takeFailures);
-                        }
-                    }
-                });
-                clearThreads[i] = Task.Run(() =>
-                {
-                    for (int j = 0; j < itemsPerThread; j++)
-                    {
-                        try
-                        {
-                            bag.Clear();
-                        }
-                        catch
-                        {
-                            Interlocked.Increment(ref clearFailures);
-                        }
-                    }
-                });
-                arrayThreads[i] = Task.Run(() =>
-                {
-                    for (int j = 0; j < itemsPerThread; j++)
-                    {
-                        try
-                        {
-                            bag.ToArray();
-                        }
-                        catch
-                        {
-                            Interlocked.Increment(ref arrayFailures);
-                        }
-                    }
-                });
-            }
-
-            Task.WaitAll(addThreads);
-            Task.WaitAll(peekThreads);
-            Task.WaitAll(takeThreads);
-            Task.WaitAll(clearThreads);
-            Task.WaitAll(arrayThreads);
-
-            Assert.Equal(0, addFailures);
-            Assert.Equal(0, peekFailures);
-            Assert.Equal(0, takeFailures);
-            Assert.Equal(0, clearFailures);
-            Assert.Equal(0, arrayFailures);
         }
 
         #region Helper Methods / Classes
